+++ ui/components/progress-indicator/base/index.stories.js
@@ -4,19 +4,19 @@ import { getDisplayElementById } from '../../../shared/helpers';
 
 import '../../../index.scss';
 
-const stories = storiesOf('Components/Rich Text Editor/Base', module).add(
+const stories = storiesOf('Components/Progress Indicator/Base', module).add(
   'Base',
   () => BaseExamples.default
 );
 
-BaseExamples.states.forEach(example => {
+BaseExamples.examples.map(example => {
   stories.add(example.label, () =>
-    getDisplayElementById(BaseExamples.states, example.id)
+    getDisplayElementById(BaseExamples.examples, example.id)
   );
 });
 
-BaseExamples.examples.forEach(example => {
+BaseExamples.states.map(example => {
   stories.add(example.label, () =>
-    getDisplayElementById(BaseExamples.examples, example.id)
+    getDisplayElementById(BaseExamples.states, example.id)
   );
 });